Improve integrationtests and unit tests (#39)
* Fix for bug GH-37

* Improvements to the author integration tests

[GH-34]

Refactored the test object to make it generic. Added a couple of traits
that shall be implemented by objects that target a particular resource.
This way, the main test object is generic and can be used by any
particular object that implements tests for a particular resource.

* Add social profiles to the author patch test

[GH-34]

* Refactored API test builders

[GH-34]

* Refactored integration tests for the /ingredient resource of the API

[GH-34]

* Add unittest to /author GET

[GH-32]

* Refactored integration tests for the /ingredient resource of the API

[GH-34]

* Refactored search tests for ingredient and author resources

[GH-34]

* Increased coverage for /author

[GH-34]

* Improved /token/request integrationt tests

[GH-34]

Added integration test to cover the validation step.